python-samples/requirements.txt
semantic_kernel
pyyaml
pydantic
jinja2
pybars3
